+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New:
 - multipointed translations for S5, including actions
 - `instance Optimizable` for symbolic structures
 - provide `whereViaBdd` also in `SMCDEL.Symbolic.S5_CUDD`
+- integration tests for web interface
 
 Changed: